backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mammoth==1.6.0
Werkzeug==2.3.7
python-docx==1.1.0